requirements.txt
crawl4ai>=0.6.0
google-genai>=1.0.0
httpx>=0.27.0
streamlit>=1.30.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
pydantic>=2.0.0
python-dotenv>=1.0.0
